+++ tests/e2e/game/unkillable-ancient.test.js
@@ -47,7 +47,7 @@ describe("P - Game with an ancient who survives from 3 werewolves attacks", () =
         chai.request(app)
             .post("/games")
             .set({ Authorization: `Bearer ${token}` })
-            .send({ players, options: { roles: { sheriff: { enabled: false } } } })
+            .send({ players, options: { roles: { sheriff: { isEnabled: false } } } })
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;